requirements.txt
blurhash==1.1.4
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.1
idna==3.10
instaloader==4.14.1
Mastodon.py==2.0.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
requests==2.32.3
six==1.17.0
urllib3==2.3.0
